I forgot that CodeForces only allows to submit 1 file, thus I merged them all into one. Reading from stdio if file is not specified
